Added initial try at some base code for the maze board
Will add general description of how this works when I have the chance, but it's fairly straightforward (I hope).

At this point, the Board class can take a 2D array of Tile objects (superclass with Floor, Wall, etc. subclasses, as shown in Main method at top) and do a couple things (place keys, set traps).

So, it'd be possible to procedurally generate a Tile 2D array, then create a Board object out of it and prep it for the game (at least to a very simple extent, atm).

LMK if I should just push the entire visual studio project, so it's easier to work out of.