Added Astar testing class with search method
This is ENTIRELY UNTESTED, and CANNOT currently be implemented into the Enemy file as-is, which is why I'm pushing it as a separate code.  The different components themselves are untested, which means the MinPQ, Comparer stuff, hashcode functions, Astar search method, etc. should be individually tested for proper functionality.  I will start debugging them later today.