Deleted old code and exported new project
There are a lot of changes that happened here, but not as many as it seems.  The project was rebuilt, so all the old files got replaced.  Alongside that, there are a lot of files that Unity generates that I decided to put back on the github, because it's really annoying to try and get this project to work on another computer without all of those files.

The most important changes are in the script files.  Old code written by other people have been removed and revamped by myself, specifically relating to Enemy.cs.