Added new features and changed game logic
Keeping the spirit of a maze-survival game, the game is shifting to be a "run away from the enemy until you get enough resources to kill him" game.  As such, the player will need to explore as much of the maze as possible, clearing as much of the floor as possible.

New Feature: Floor Score - the more explored, the more a bar will fill up.  The player can use the R keybind to stun the enemy for a duration relative to how much of the bar they have full.  If the bar is completely full and they use it, then they'll kill the enemy.