Added one trap architype
A pain trap has been added, causing 1 health loss to the player when they walk over it.  This occurs 2.5% of the time.

The ice traps have proved to be a bit more tricky to set up, due to the turn system.  The outline has been made, but it isn't implemented completely yet.